+++ xarray/coding/cftime_offsets.py
@@ -41,25 +41,35 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 import re
-from datetime import timedelta
+from datetime import datetime, timedelta
 from distutils.version import LooseVersion
 from functools import partial
 from typing import ClassVar, Optional
 
 import numpy as np
+import pandas as pd
 
+from ..core.common import is_np_datetime_like
 from ..core.pdcompat import count_not_none
 from .cftimeindex import CFTimeIndex, _parse_iso8601_with_reso
-from .times import format_cftime_datetime
+from .times import (
+    _is_numpy_compatible_time_range,
+    _is_standard_calendar,
+    convert_time_or_go_back,
+    format_cftime_datetime,
+)
 
 
-def get_date_type(calendar):
+def get_date_type(calendar, use_cftime=True):
     """Return the cftime date type for a given calendar name."""
     try:
         import cftime
     except ImportError:
         raise ImportError("cftime is required for dates with non-standard calendars")
     else:
+        if _is_standard_calendar(calendar) and not use_cftime:
+            return pd.Timestamp
+
         calendars = {
             "noleap": cftime.DatetimeNoLeap,
             "360_day": cftime.Datetime360Day,
@@ -708,6 +718,8 @@ def to_cftime_datetime(date_str_or_date, calendar=None):
         return date
     elif isinstance(date_str_or_date, cftime.datetime):
         return date_str_or_date
+    elif isinstance(date_str_or_date, (datetime, pd.Timestamp)):
+        return cftime.DatetimeProlepticGregorian(*date_str_or_date.timetuple())
     else:
         raise TypeError(
             "date_str_or_date must be a string or a "
@@ -1016,3 +1028,181 @@ def cftime_range(
         dates = dates[:-1]
 
     return CFTimeIndex(dates, name=name)
+
+
+def date_range(
+    start=None,
+    end=None,
+    periods=None,
+    freq="D",
+    tz=None,
+    normalize=False,
+    name=None,
+    closed=None,
+    calendar="standard",
+    use_cftime=None,
+):
+    """Return a fixed frequency datetime index.
+
+    The type (CFTimeIndex or pd.DatetimeIndex) of the returned index depends
+    on the requested calendar and on `use_cftime`.
+
+    Parameters
+    ----------
+    start : str or datetime-like, optional
+        Left bound for generating dates.
+    end : str or datetime-like, optional
+        Right bound for generating dates.
+    periods : int, optional
+        Number of periods to generate.
+    freq : str or None, default: "D"
+        Frequency strings can have multiples, e.g. "5H".
+    tz : str or tzinfo, optional
+        Time zone name for returning localized DatetimeIndex, for example
+        'Asia/Hong_Kong'. By default, the resulting DatetimeIndex is
+        timezone-naive. Only valid with pandas DatetimeIndex.
+    normalize : bool, default: False
+        Normalize start/end dates to midnight before generating date range.
+    name : str, default: None
+        Name of the resulting index
+    closed : {"left", "right"} or None, default: None
+        Make the interval closed with respect to the given frequency to the
+        "left", "right", or both sides (None).
+    calendar : str, default: "standard"
+        Calendar type for the datetimes.
+    use_cftime : boolean, optional
+        If True, always return a CFTimeIndex.
+        If False, return a pd.DatetimeIndex if possible or raise a ValueError.
+        If None (default), return a pd.DatetimeIndex if possible, otherwise return a CFTimeIndex.
+        Defaults to False if `tz` is not None.
+
+    Returns
+    -------
+    CFTimeIndex or pd.DatetimeIndex
+
+    See also
+    --------
+    pandas.date_range
+    cftime_range
+    """
+    from .times import _is_standard_calendar
+
+    if tz is not None:
+        use_cftime = False
+
+    if _is_standard_calendar(calendar) and use_cftime is not True:
+        try:
+            return pd.date_range(
+                start=start,
+                end=end,
+                periods=periods,
+                freq=freq,
+                tz=tz,
+                normalize=normalize,
+                name=name,
+                closed=closed,
+            )
+        except pd.errors.OutOfBoundsDatetime as err:
+            if use_cftime is False:
+                raise ValueError(
+                    "Date range is invalid for pandas DatetimeIndex, try using `use_cftime=True`."
+                ) from err
+    elif use_cftime is False:
+        raise ValueError(
+            f"Invalid calendar {calendar} for pandas DatetimeIndex, try using `use_cftime=True`."
+        )
+
+    return cftime_range(
+        start=start,
+        end=end,
+        periods=periods,
+        freq=freq,
+        normalize=normalize,
+        name=name,
+        closed=closed,
+        calendar=calendar,
+    )
+
+
+def date_range_like(source, calendar, use_cftime=None):
+    """Generate a datetime array with the same frequency, start and end as another one, but in a different calendar.
+
+    Parameters
+    ----------
+    source : DataArray or CFTimeIndex or pd.DatetimeIndex
+      1D datetime array
+    calendar : str
+      New calendar name.
+    use_cftime : bool, optional
+      If True, the output uses cftime objects. If None (default), numpy objects are used if possible.
+      If False, numpy objects are used or an error is raised.
+
+    Returns
+    -------
+    DataArray
+      1D datetime coordinate with the same start, end and frequency as the source, but in the new calendar.
+      The start date is assumed to exist in the target calendar.
+      If the end date doesn't exist, the code tries 1 and 2 calendar days before.
+      Exception when the source is daily or coarser, then if the end of the input range is on
+      the last day of the month, the output range will also end on the last day of the month in the new calendar.
+    """
+    from .frequencies import infer_freq
+
+    freq = infer_freq(source)
+    if freq is None:
+        raise ValueError(
+            "`date_range_like` was unable to generate a range as the source frequency was not inferrable."
+        )
+
+    # Arguments Checks for target
+    if use_cftime is not True:
+        if _is_standard_calendar(calendar):
+            if _is_numpy_compatible_time_range(source):
+                # Conversion is possible with pandas, force False if it was None
+                use_cftime = False
+            elif use_cftime is False:
+                raise ValueError(
+                    "Source time range is not valid for numpy datetimes. Try using `use_cftime=True`."
+                )
+            # else : Default to cftime
+        elif use_cftime is False:
+            # target calendar is ctime-only.
+            raise ValueError(
+                f"Calendar '{calendar}' is only valid with cftime. Try using `use_cftime=True`."
+            )
+        else:
+            use_cftime = True
+
+    src_start = source.values.min()
+    src_end = source.values.max()
+    if is_np_datetime_like(source.dtype):
+        src_cal = "default"
+        # We want to use datetime fields (datetime64 object don't have them)
+        src_start = pd.Timestamp(src_start)
+        src_end = pd.Timestamp(src_end)
+    else:
+        if isinstance(source, CFTimeIndex):
+            src_cal = source.calendar
+        else:  # DataArray
+            src_cal = source.dt.calendar
+
+    tgt_cal = calendar if use_cftime else "default"
+    if src_cal == tgt_cal:
+        return source
+
+    date_type = get_date_type(calendar, use_cftime)
+    start = convert_time_or_go_back(src_start, date_type)
+    end = convert_time_or_go_back(src_end, date_type)
+
+    # For the cases where the source ends on the end of the month, we expect the same in the new calendar.
+    if src_end.day == src_end.daysinmonth and isinstance(
+        to_offset(freq), (YearEnd, QuarterEnd, MonthEnd, Day)
+    ):
+        end = end.replace(day=end.daysinmonth)
+
+    return date_range(
+        start=start.isoformat(),
+        end=end.isoformat(),
+        freq=freq,
+        calendar=calendar,
+    )